add redis module

+++ src/h2d_lua_api.h
@@ -11,7 +11,7 @@ int h2d_lua_api_thread_resume(lua_State *L, struct h2d_request *r);
 void h2d_lua_api_thread_free(lua_State *L);
 
 const char *h2d_lua_api_call_lstring(struct h2d_request *r,
-		wuy_cflua_function_t f, size_t *plen);
+		wuy_cflua_function_t f, int *plen);
 
 int h2d_lua_api_call_boolean(struct h2d_request *r,
 		wuy_cflua_function_t f);
