+++ src/modules/h2d_upstream_hash.c
@@ -83,7 +83,7 @@ static struct h2d_upstream_address *h2d_upstream_hash_pick(
 {
 	struct h2d_upstream_hash_conf *conf = upstream->lb_confs[h2d_upstream_hash.index];
 
-	size_t key_len;
+	int key_len;
 	const char *key_str = h2d_lua_api_call_lstring(r, conf->key, &key_len);
 	if (key_str == NULL) {
 		return NULL;
